+++ mobile/android/android-components/components/support/ktx/src/test/java/mozilla/components/support/ktx/android/util/DisplayMetricsTest.kt
@@ -6,22 +6,24 @@ package mozilla.components.support.ktx.android.util
 
 import android.util.DisplayMetrics
 import androidx.test.ext.junit.runners.AndroidJUnit4
-import mozilla.components.support.test.robolectric.testContext
-import org.junit.Assert
+import org.junit.Assert.assertEquals
 import org.junit.Test
 import org.junit.runner.RunWith
+import kotlin.math.roundToInt
 
 @RunWith(AndroidJUnit4::class)
 class DisplayMetricsTest {
 
     @Test
     fun `dp returns same value as manual conversion`() {
-        val metrics = testContext.resources.displayMetrics
+        val metrics = DisplayMetrics().apply {
+            density = 2.75f
+        }
 
-        for (i in 1..500) {
-            val px = (i * (metrics.densityDpi.toFloat() / DisplayMetrics.DENSITY_DEFAULT)).toInt()
-            Assert.assertEquals(px, i.dpToPx(metrics))
-            Assert.assertNotEquals(0, i.dpToPx(metrics))
+        for (i in 1..10) {
+            val expected = (i * 2.75).roundToInt()
+
+            assertEquals(expected, i.dpToPx(metrics))
         }
     }
 }