+++ mobile/android/android-components/docs/changelog.md
@@ -4,7 +4,9 @@ title: Changelog
 permalink: /changelog/
 ---
 
-# 132.0 (In Development)
+# 133.0 (In Development)
+
+# 132.0
 * **feature-awesomebar**
   * The `onCancelEditing` now returns a result based on the `onStartEditing` and `onStopEditing` callback. [Bug 1917496](https://bugzilla.mozilla.org/show_bug.cgi?id=1917496)